+++ packages/core/src/rx/operators/creation/index.ts
@@ -7,4 +7,7 @@ export { default as timer } from "./timer";
 export { default as throwError } from "./throwError";
 export { default as fromDOMMutations } from "./from-dom-mutations";
 export { default as of } from "./of";
-export { default as from } from "./from";
+export {
+  default as from, StreamInput, StreamInterop, StreamInteropRxJS,
+  ObserverDefinitionInterop
+} from "./from";