+++ packages/core/src/rx/operators/creation/throwError.ts
@@ -1,7 +1,16 @@
 import Stream from "../../stream";
 
 /**
-   TODO
+   Creates a new stream that immediatly emits an error value. Generally speaking, you can
+   just `throw` an error in a Stream and it will be picked up like you would expect. There
+   are situations however where you may want to create a Stream with this operator, which
+   will emit the error value on every subscription.
+
+   @example
+   ```typescript
+   rx.throwError(new Error('hi')).subscribe({ error: console.log });
+   // prints the error
+   ```
 
    @group Stream Creation Operators
  */